+++ packages/expo-router/src/getLinkingConfig.ts
@@ -5,7 +5,7 @@ import { INTERNAL_SLOT_NAME, NOT_FOUND_ROUTE_NAME, SITEMAP_ROUTE_NAME } from './
 import type { State } from './fork/getPathFromState';
 import { getReactNavigationConfig } from './getReactNavigationConfig';
 import { applyRedirects } from './getRoutesRedirects';
-import type { StoreRedirects } from './global-state/router-store';
+import type { StoreRedirects } from './global-state/types';
 import { getInitialURL, getPathFromState, getStateFromPath, subscribe } from './link/linking';
 import type { LinkingOptions } from './react-navigation/native';
 import type { NativeIntent, RequireContext } from './types';